requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
dj-database-url==2.0.0
Django==4.2.3
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.6
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
